Search bar toggle implemented
